refactor: split up tmath.cpp into units
