+++ .gitignore
@@ -232,3 +232,7 @@ $RECYCLE.BIN/
 
 # Windows shortcuts
 *.lnk
+
+build/*
+objects/*
+lib/*
