+++ src/abs.cpp
@@ -1,0 +1,6 @@
+#include "tmath.hpp"
+
+TMath::DOUBLE TMath::abs(const DOUBLE &x) {
+	if (x < 0) return -x;
+	else return x;
+}
